bugfix circular import
